Fix formatting of the list, adjust presentation to be clear(er) and more factual
I removed a note about "Nothing in the real world is [linear]!" because it is
misleading.  Many "things" are linear, non-0 error function could be due to
measurement errors, etc.  So I decided to avoid this comment.

I used more of :math: mode to get equations not hard to see among the text, etc