Minor word edits
## Problem/Motivation
Under the section: How the interpreter looks up attributes
The word 'there' was underlined, but didn't render properly in the browser line 163. 
The Grandparent class referencing on line 163 was a little awkward.

## Proposed resolution
Remove the word 'there' and change the way grandparent class is referenced.

## Remaining tasks
Decide if these changes make sense.