Clarifying instructions
Proposing ideas to address two common misinterpretations that students have with this problem:

(1) that last_four should take in a list of numbers (rather than a single number) and return a list of the last four digits of each number. E.g. last_four takes in [17573005, 17572342, 17579000] and returns [3005, 2342, 9000]

(2) that sorting the list of ids should take place within the definition of last_four (rather than outside of last_four)

Although the original instructions seem clear, it still might be a good idea to add more clarifying language to prevent future confusion/errors